pyproject.toml
[project]
name = "rustsensei"
version = "0.1.0"
description = "Apprendre le Rust en français, localement"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Abdel TOUATI", email = "abdel.touati@gmail.com" }
]
keywords = ["rust", "education", "local", "rag"]

dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "huggingface-hub>=0.20.0",
    # RAG (versions récentes pour Python 3.13)
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=3.0.0",
    # Data
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]
eval = [
    "nltk>=3.8.0",
]

[project.scripts]
rustsensei = "app.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
